move handoff to own component for suspense loading
